fix(surveys): apply display status + relocate shared types per review
- committee-overview: switch openSurveysCount/pendingSurveys from
  getEffectiveSurveyStatus to getSurveyDisplayStatus so the API's
  response_status='closed' sentinel is honored (matches the dashboard fix)
- shared: move COMBINED_SURVEY_STATUS and CombinedSurveyStatus from
  survey.utils.ts to survey.constants.ts; lookup constants belong in
  the constants subpath per shared-package conventions
- shared: move SurveyStatusInput and SurveyDisplayStatusInput from
  survey.utils.ts to survey.interface.ts so they're discoverable via
  @lfx-one/shared/interfaces
- update pipe and component imports to source types/constants from the
  appropriate @lfx-one/shared subpath

Made-with: Cursor
