+++ apps/lfx-one/src/app/modules/my-activity/components/surveys-table/surveys-table.component.ts
@@ -12,8 +12,9 @@ import { SelectComponent } from '@components/select/select.component';
 import { TableComponent } from '@components/table/table.component';
 import { TagComponent } from '@components/tag/tag.component';
 import { COMBINED_SURVEY_STATUS_LABELS, MY_ACTIVITY_FILTER_LABELS } from '@lfx-one/shared';
+import { CombinedSurveyStatus } from '@lfx-one/shared/constants';
 import { UserSurvey } from '@lfx-one/shared/interfaces';
-import { CombinedSurveyStatus, getCombinedSurveyStatus } from '@lfx-one/shared/utils';
+import { getCombinedSurveyStatus } from '@lfx-one/shared/utils';
 import { CanTakeSurveyPipe } from '@pipes/can-take-survey.pipe';
 import { CombinedSurveyStatusLabelPipe } from '@pipes/combined-survey-status-label.pipe';
 import { CombinedSurveyStatusSeverityPipe } from '@pipes/combined-survey-status-severity.pipe';
